Add basic user profile details
