+++ src/skills/skills.ts
@@ -1,5 +1,7 @@
 import { undef, union } from "typescript-json-decoder";
 
+import { unique } from "~/src/utils";
+
 /** All available skill levels we work with */
 export const SKILL_LEVELS = ["junior", "medior", "senior", "mentor"] as const;
 
@@ -138,3 +140,16 @@ export function decodeSkillSelection(input: string): SkillSelection {
       .reduce(addSkill, {})
   );
 }
+
+export const skillsToHashtags = (skills: string) => {
+  const uppercaseFirst = (s: string) =>
+    s.charAt(0).toLocaleUpperCase() + s.slice(1);
+  const categories = skills
+    .split(/;\s*/)
+    .map((s) => s.split(/\s*\/\s*/).shift())
+    .map((c) => c?.split(" ").map(uppercaseFirst).join(""));
+  return unique(categories)
+    .sort()
+    .map((c) => "#" + c)
+    .join(" ");
+};
